Address some Visual C++ compiler warnings

+++ src/ox/std/strops.cpp
@@ -34,7 +34,7 @@ int atoi(const char *str) {
 	int total = 0;
 	int multiplier = 1;
 
-	for (int i = strlen(str) - 1; i >= 0; i--) {
+	for (size_t i = strlen(str) - 1; i >= 0; i--) {
 		total += (str[i] - '0') * multiplier;
 		multiplier *= 10;
 	}
